x64: brgemm: fix over-widened brgemm_desc_t/brgemm_attr_t field types
The initial infra widening (widen brgemm_desc_t/brgemm_attr_t fields to
dim_t) was over-broad: it widened many bounded, hardware/ISA-scale
code-generation values (register/tile counts, blocking hints, block
SIZES, remainders, type sizes, VNNI step constants) to dim_t even
though they never scale with tensor size. Only genuinely tensor-scale
values (leading dimensions, block COUNTS such as bdb/ldb/rdb and
bdb2/ldb2, buffer sizes) should remain dim_t.

Reclassify back to int: max_top_vpad/max_bottom_vpad/max_top_bpad/
max_bottom_bpad (bounded by MAX_VPAD), bd_mask_level, hint_bs_group/
hint_bd_block/hint_ld_block/hint_bd_block2/hint_ld_block2, bs_group,
bd_block/bd_block2/ld_block/ld_block2/rd_block and their *_tail
remainders, rd_step/ld_step, typesize_A/B/C/D/bias, is_M_tail,
gemv_transa_bd_unroll/gemv_tail and the gemv_bd_block()/gemv_bdb_tail()/
gemv_num_acc_blocks() accessors, and the AMX tile-register decomposition
getters (get_bd_block2()/get_ld_block2()/get_num_C_tiles()/
get_C_tensor()/get_num_A_tiles()/get_A_tensor()/get_num_B_tiles()/
get_B_tensor()), all bounded by AMX_TILES_NUM.

Propagate these type corrections through every cascading call site in
brgemm.cpp, brgemm_utils.cpp, jit_brgemm_amx_uker.cpp, and
jit_brgemm_kernel.cpp.

Also fix brgemm_containers.hpp/.cpp: operator[], insert(), and
maybe_tile_configure() took/returned dim_t indices while every caller
in the tree (jit_brgemm_conv.cpp, jit_brgemm_conv_bwd_strided.cpp, ...)
passes int - a non-const dim_t& parameter cannot bind to an int
argument, so this was a pre-existing hard compile error blocking
brgemm_convolution_fwd_t template instantiation entirely. Reclassify
the container index type back to int to match every real caller.

Verified via -fsyntax-only -Wshorten-64-to-32 -Wsign-compare scans
(with -ferror-limit=0 for an apples-to-apples before/after diff): zero
new warnings introduced in any touched or downstream file; the
apparent warning-count increase in jit_brgemm_conv.cpp/
jit_brgemm_inner_product.cpp/brgemm_matmul.cpp between before/after
scans was solely due to the pre-existing hard error truncating the
syntax-only scan early, not new warnings. clang-tidy clean on all
touched files.

+++ src/cpu/x64/brgemm/brgemm.cpp
@@ -348,7 +348,8 @@ status_t brgemm_desc_set_postops(brgemm_desc_t *brg,
     brg->dt_bias = dt_bias;
     brg->typesize_bias = (dt_bias == data_type::undef)
             ? 0
-            : types::data_type_size(brg->dt_bias);
+            // Type sizes are always 1/2/4/8, safe to narrow.
+            : static_cast<int>(types::data_type_size(brg->dt_bias));
 
     brg->LDD = LDD;
     brg->is_runtime_ldd = is_runtime_value(LDD);
@@ -424,7 +425,8 @@ status_t brgemm_desc_set_postops(brgemm_desc_t *brg,
         return status::unimplemented;
 
     brg->dt_d = dt_d;
-    brg->typesize_D = types::data_type_size(brg->dt_d);
+    // Type sizes are always 1/2/4/8, safe to narrow.
+    brg->typesize_D = static_cast<int>(types::data_type_size(brg->dt_d));
 
     if (!IMPLICATION(brg->is_int8 && brg->dt_d == bf16,
                 is_superset(brg->isa_impl, avx512_core)
